feat/PRSD-724 review mark ups
